diffusion: drop .float() in CE; restore Spark batch_size=128
F.cross_entropy upcasts to fp32 internally for log_softmax stability but
doesn't retain the fp32 copy in the autograd graph. The explicit
`.float()` cast we had was materializing a B*T*V fp32 tensor + its
gradient — at B=128, V=50304 that was ~52 GB unnecessarily, the actual
cause of the Spark OOM.

Skipping the cast: peak training memory drops from ~115 GB to ~63 GB,
batch=128 fits cleanly. Numerical impact at our scale is negligible
(every production-scale LM training does CE on bf16 logits).

Restores Spark sweep configs to batch_size=128, grad_accum=1.

Smoke test still passes.

+++ nanodiff/diffusion.py
@@ -52,8 +52,14 @@ def diffusion_loss(logits, x0, mask, t):
         scalar loss (a stochastic upper bound on the negative log-likelihood).
     """
     B, T, V = logits.shape
+    # NOTE: we pass bf16 logits straight to cross_entropy — `F.cross_entropy`
+    # already upcasts to fp32 *internally* for log_softmax (numerical stability)
+    # but does not store an fp32 copy in the autograd graph. Skipping `.float()`
+    # here saves ~B*T*V*4 bytes for the cast + ~same for its gradient — at
+    # B=128, T=1024, V=50304 that's ~52 GB. Critical for fitting big-batch
+    # training in a single-GPU box's memory.
     ce = F.cross_entropy(
-        logits.reshape(-1, V).float(),
+        logits.reshape(-1, V),
         x0.reshape(-1),
         reduction="none",
     ).reshape(B, T)
